CYTOSCAPE-12898.  Add note that "fill color" isn't used for images, remove fillColor and fillOpacity from return values, and change the image opacity to go from 0-100 rather than 0-1.

+++ ding-impl/ding-presentation-impl/src/main/java/org/cytoscape/ding/impl/cyannotator/tasks/ImageAnnotationTunables.java
@@ -31,7 +31,7 @@ public class ImageAnnotationTunables extends AbstractAnnotationTunables {
 
   @Tunable(context="nogui", 
            description="The image opacity adjustment")
-  public Integer opacity = null;
+  public Double opacity = null;
 
   /*
    * FIXME: How do we handle SVG?
@@ -64,6 +64,6 @@ public class ImageAnnotationTunables extends AbstractAnnotationTunables {
     }
     if (contrast != null) iAnnotation.setImageContrast(contrast);
     if (brightness != null) iAnnotation.setImageBrightness(brightness);
-    if (opacity != null) iAnnotation.setImageOpacity(opacity);
+    if (opacity != null) iAnnotation.setImageOpacity(opacity.floatValue());
   }
 }
